Removed unecessary logic and renamed view
